Refs #5 - Did not work in jetty, because !SimpleUrlHandlerMapping does not sort mapping keys, the most suitable mapping is decided according to the longest match, so the static content will be matched by other controllers.
SUHM uses hashmap to store mappings, which does not sort them according to priority. Instead, the solution is to define another SUHM and give it an order property, and define higher priority mappings in there. 

So at the end, we first check if the url is static, and forward to default servlet. On the contrary, in the default spring annotation tradition, spring first checks the @request mappings and forwards to default servlet if there is no match. I think my solution is better :)

Also removed pageContext prefixes from jsps, for simplicity and clarity. Instead /[path]/edit redirects to /admin/editPage?path=[path]