Refs #10 - Ajax review changes, attribute history, view changes (diff) implemented. In attribute history dialog (old revert dialog) user can diff any version to any other version. In review changes, changed attributes are detected and user can view changes (diff), revert to original state. For diff using prettydiff javascript prettifier and diff engine. Requires single js (yay!) and for the style I reduced it to the only necessary parts. Added mapping to enter to admin pages by directly typing /admin.
''The hardest part is to come up with a work model, what to call dialogs, what to call buttons, etc. I cannot decide what to call to buttons in attr history (was revert...) revert/publish or preview/revert. What will be intuitive and straightforward? May be this is because I'm not sure what the target audience is, too naive or too expert. Will they get it when they see 'revert'? Indeed, it's hard.''

Also decided not to implement page history feature. It's hard to implement, and the javascript model will have all history of all attributes of the page, too much data. Reverting to a specific date page state would be nice, may be it can be implemented later if the users demand it. Moreover it can be confused with attribute history. Want to keep it simple at first.