Refs #12 - Dynamic DOM flaw is fixed. Action dialog implemented, can edit and save the template.
Dynamic DOM flaw took the whole week to overcome. The problem is that we cannot rely on the browser generated html, since the scripts heavily modify it, adds nodes, classes, etc. When the template html is overwritten with browser reported inner html of the document, things screw up badly. There was not an easy fix, how could we detect which nodes are script generated? Disabling javascript was not an option since most templates use it for slideshows and layout. Then thought of adding custom attributes to all DOM elements in the template html and tracking changes. First tried to do it using javascript but there is no javascript html parsing library other than the browser itself. We needed a tree update and merge solution. Since the user can update any sub-tree of the html, the modified sub trees had to be renumbered with the custom attribute. After coming up with the tree update algorithm and failing to implement with javascript (first I tried to cache inner html of all nodes recursively, but this failed on sub-tree updates). Finally, at the end of the week figured out that if we keep state on the server only, and cache JSoup object things will be easy and straightforward.

In the current implementation; on edit html dialog we get the actual inner html of the selected node (by custom id), on update send the updated html to the server, server merges it with the state, regenerates sub-tree custom attributes, and javascript updates the element inner html with this renumbered tree.

Other changes;

 * Template editing iframe is retrieved from session (modified html is put to session on editTemplate page)
 * Iframe is added with javascript (otherwise browser loads it twice)
 * getTemplateHtml, updateTemplateHtml are called on edit html dialog, modifies the state
 * saveTemplate saves the current state to the template html (overwrites, no versioning yet)
 * Action dialog pops up whenever an element is selected via tree or inspection
 * Inspection skips dom generated elements, firebug lite top bar fbTop is hidden with direct html
 
Review changes and view changes (versioning) not implemented yet.