Fixes #12 - Dynamically edit template is OK. Process template is implemented to fix respath problem. Processed template html includes absolute paths for links, css and imports. Imported css's are also updated on initial processing (on addTemplate). Original resources are left intact (except css's which I had to modify).
Page paths can include dot and slash. Users can also access static htm/html resources. If a page with path.html cannot be found it is retreated as static. This helps us to use static html which are not made template. This means user can use only resource feature to host a static site.

fmgcms-id attributes are not displayed in inspector.