Fixes #14 - Bugs are fixed. Jsoup formatting problem automatically fixed after upgrading to jsoup 1.6.3 version. For aloha, changed plugin file (fmg-plugin) dependencies to be loaded from itself (jquery ui script and css), and removed jqueryui import from inject editor. Also plugin uses aloha jquery instead of global jquery $. Changed the plugin messages to English.
NOTE: Tried to update to Aloha new version, but they made a lot of changes, postponing (besides I don't like the new UI).

Another possbile bug also postponed: Template links are processed, but if the user selects an image as attribute, the links are retrieved from the original html, so they're not processed and links are broken. May need to process page attributes (html fragment) as in template html.