Refs #18 - Seamless UX scenario init.
Domain independent signup and login. User is auto logged in to his primary domain if not authorized to the domain he initially logs in to. And can switch to any site he owns without relogin. The impl detail is we do a checkLogin to the target domain to prevent generating authToken, so JSONP is used to overcome single origin policy.

Also the login does not add user to session anymore, just issues an authToken, which can be consumed at any domain (at interceptor).

Username is changed to lowercase on signup.

Signup UX is updated. Changed wording. No need to specify a domain. No need to login after signup. Sample designs are copied to the user after signup.

Delete account feature is implemented. All sites, authors, resources are removed from the system.

Not found error page can be added as 404.html static or a page. No html is returned if none exists.