Refs #18 - Typed attribute system implemented. Default values for attributes and attribute enumerations are kept at !TemplateAttribute. Attributes are not scanned on every occasion, anymore. When creating a page they are added from tA. When updated at editTemplate they are written to tA. Even removed attributes are kept as removed flag set. Then when they are re-added pages can use already existing attrs.
Nested attr creation is blocked by checking them on makeAttribute.

The blocking bug is also fixed (#15). When attributes are created they are processed as in template and the resource relative links are absolutized. 

More testing is required for these changes, and it can be further improved.

Template can be duplicated from templates page.

User voice and google analytics scripts are moved to the bottom.